+++ R/function_logging.R
@@ -7,6 +7,15 @@
 
 #' @title Update giotto parameters
 #' @name update_giotto_params
+#' @description
+#' For developer use. Adds an entry to the `giotto` object object history.
+#' Care currently needs to be taken when a function that contains a call to
+#' this function is called from within yet another function. In such cases,
+#' a `toplevel < 0` or setting a temporary `"giotto.update_param" = FALSE` with
+#' `GiottoUtils::gwith_option()` may be the best option to avoid either
+#' evaluation errors or strange history entries. A new `update_giotto_params()`
+#' call can then be added that describes the function of the topmost function
+#' if desired.
 #' @param gobject giotto object
 #' @param description description of function run
 #' @param return_gobject logical. Whether the giotto object should be returned
@@ -22,7 +31,7 @@ update_giotto_params <- function(gobject,
     description = "_test",
     return_gobject = TRUE,
     toplevel = 2) {
-    if (toplevel < 0) {
+    if (toplevel < 0 || !getOption("giotto.update_param", TRUE)) {
         return(gobject)
     } # skip if toplevel negative