+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - `spatUnit()<-` and `featType()<-` `list` methods
 - `set_default_spat_unit()` and `set_default_feat_type()` now look for defaults when given `NA_character_` inputs as well.
+- `update_giotto_params()` can now be turned off with `options("giotto.update_param" = FALSE)`
 
 # GiottoClass 0.4.4 (2024/11/14)